chore: move tailwind util files for colors, opacities, and shadows up
